+++ lib/helpers/dependencies_injection.dart
@@ -1,24 +1,32 @@
 /// lib/helpers/dependencies_injection.dart
 ///
 /// dependencies registration and injection
+import 'package:dio/dio.dart';
 import 'package:get_it/get_it.dart';
 import 'package:logger/logger.dart';
 import 'package:ourjourneys/helpers/get_platform_service.dart';
+import 'package:ourjourneys/services/api/api_service.dart';
 import 'package:ourjourneys/services/auth/acc/auth_service.dart';
 import 'package:ourjourneys/services/auth/acc/auth_wrapper.dart';
 import 'package:ourjourneys/services/auth/local/local_auth_service.dart';
+import 'package:ourjourneys/services/cloud/cloud_file_service.dart';
 import 'package:ourjourneys/services/configs/utils/permission_service.dart';
+import 'package:ourjourneys/services/network/dio_handler.dart';
 import 'package:ourjourneys/services/object_storage/cloud_object_storage_service.dart';
 import 'package:ourjourneys/services/object_storage/cloud_object_storage_wrapper.dart';
 import 'package:ourjourneys/services/package/package_info_provider.dart';
 import 'package:ourjourneys/services/package/package_info_service.dart';
 import 'package:ourjourneys/services/pref/shared_pref_service.dart';
+import 'package:pretty_dio_logger/pretty_dio_logger.dart';
 
 final GetIt getIt = GetIt.instance;
 
 Future<void> setupDependencies() async {
   setupAuthServices();
   setupCloudObjectStorageServices();
+  setupDio();
+  setupApiServices();
+  setupCloudFileServices();
   // getIt.registerLazySingleton(() => FirestoreService());
   // getIt.registerLazySingleton(() => FirestoreWrapper());
   getIt.registerLazySingleton(() => PackageInfoService());
@@ -47,6 +55,19 @@ void setupAuthServices() {
   getIt.registerLazySingleton(() => AuthWrapper());
 }
 
-void setupLogger() {
+void setupLoggers() {
   getIt.registerSingleton<Logger>(Logger());
+  getIt.registerSingleton<PrettyDioLogger>(PrettyDioLogger());
+}
+
+void setupApiServices() {
+  getIt.registerLazySingleton(() => ApiService());
+}
+
+void setupDio() {
+  getIt.registerLazySingleton<DioHandler>(() => DioHandler(Dio()));
+}
+
+void setupCloudFileServices() {
+  getIt.registerLazySingleton(() => CloudFileService());
 }