+++ CHANGELOG.md
@@ -38,3 +38,4 @@
 - Initial Development: Lambda functions draft (verify is working) connected to the api gateway, and other improvements on April 10, 2025 -> **0.2.8+1** by Luke
 - Initial Development: Removed support for iOS and Android temporarily on April 10, 2025 -> **0.2.8+2** by Luke
 - Initial Development: All Lambda functions (verify, gen_uploads, gen_downloads, deletes) are working on basic features, and other improvements on April 11, 2025 -> **0.2.8+3** by Luke
+- Initial Development: Enhanced deletes Lambda function, tested the ability to retrieve and upload image, and other improvements on April 12, 2025 -> **0.2.9** by Luke